demo/requirements.txt
tensorflow
numpy
Pillow
gradio
opencv-python
keras